edit/play: tempo map, wait-mode fixes, dense-chart perf, e/r bar tempo
Batch of editor + play-mode work developed on this branch.

Composer / core
- Variable per-bar tempo map (bar_starts): BPM changes re-time notes onto the
  new bar layout; per-bar nudge tools — length (grid-only) vs tempo (re-times
  notes) — plus insert/cut bar with ripple.
- Signed backing offset with a silent lead-in (audio_start_us: i64).
- Cursor note-picking fallback so off-grid imported onsets are still selectable.

Wait mode ("pause on note")
- Fresh-strike requirement: a note held over from the previous step no longer
  auto-advances a repeat of the same pitch; re-articulating any one required
  pitch suffices (generous, so a legato common tone still passes).
- Freeze at the note onset: clamp an armed tick so a frame stall can't carry the
  clock past the note and freeze after the fact.

Dense-chart performance (edit + play)
- Viewport culling: binary-search the visible time window instead of scanning
  the whole note list every frame, via a shared, unit-tested helper (cull.ts).
- Note-invariant actions emit a lightweight notes-stripped `meta` event and the
  run_action return is snapshot-free, so a cursor keystroke no longer
  re-serialises the whole chart (fixes cursor input lag on large pieces). `hand`
  added to the notes fingerprint so hand edits still count as note changes.

Keybindings
- Per-bar tempo moved off F/V (taken by the split editor / video backdrop in the
  Tauri frontend) to e/r; the four bar ops now sit on the QWER row.

Gate: fmt, workspace clippy, and all workspace tests pass; frontend tsc + vitest
(incl. new cull tests) pass.

+++ crates/import/tests/split_bundle.rs
@@ -128,7 +128,7 @@ fn split_writes_kept_parts_with_copied_media_and_derived_offsets() {
 
         let b = meta.backing.expect("backing persisted");
         assert_eq!(b.file, "backing.wav");
-        assert_eq!(b.audio_start_us, 250_000 + seg.start_us);
+        assert_eq!(b.audio_start_us, 250_000 + seg.start_us as i64);
 
         let v = meta.video.expect("video persisted");
         assert_eq!(v.file, "background.mp4");
